+++ packages/sdk/shared/src/types/index.ts
@@ -37,7 +37,6 @@ export * from './attachment';
 export * from './card';
 export * from './ipc';
 export * from './log';
-export * from './payment';
 export * from './response';
 
 export * from './apiException';